+++ ex01/Form.hpp
@@ -6,7 +6,7 @@
 /*   By: arabelo- <arabelo-@student.42lisboa.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/29 17:32:26 by arabelo-          #+#    #+#             */
-/*   Updated: 2024/07/02 22:54:57 by arabelo-         ###   ########.fr       */
+/*   Updated: 2024/08/27 18:23:44 by arabelo-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,13 +21,13 @@ class	Form {
 	private:
 		const std::string	_name;
 		bool				_isSigned;
-		const size_t		_gradeToSign;
-		const size_t		_gradeToExec;
+		const int			_gradeToSign;
+		const int			_gradeToExec;
 	public:
 		Form(void);
 		Form(const Form &obj);
 		Form	&operator=(const Form &obj);
-		Form(std::string name, size_t gradeToSign, size_t gradeToExec);
+		Form(std::string name, int gradeToSign, int gradeToExec);
 		~Form(void);
 
 		void		beSigned(Bureaucrat &obj);
@@ -35,8 +35,8 @@ class	Form {
 
 		std::string	getName(void) const;
 		bool		getIsSigned(void) const;
-		size_t		getGradeToSign(void) const;
-		size_t		getGradeToExec(void) const;
+		int			getGradeToSign(void) const;
+		int			getGradeToExec(void) const;
 
 		class	GradeTooLowException: public std::exception {
 			public:
